requirements.txt
torch
numpy
noise
pillow
matplotlib
torchvision
tqdm
kagglehub